reformulamos el navbar, usamos context
