+++ components/Login.js
@@ -1,0 +1,53 @@
+import { useAuth } from "@/context/AuthContext";
+import { Container } from "@mui/material";
+import { useRouter } from "next/router";
+import { useState } from "react";
+
+const Login = () => {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const { login } = useAuth();
+  const router = useRouter();
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    const success = await login(email, password);
+    if (success) {
+      router.push("/");
+    } else {
+      console.error("Error al logear");
+    }
+  };
+  return (
+    <Container
+      maxWidth="sm"
+      sx={{ marginTop: "100px", border: "2px solid grey", borderRadius: "5px" }}
+    >
+      <form onSubmit={handleSubmit}>
+        <div>
+          <label htmlFor="email">Email: </label>
+          <input
+            type="email"
+            id="email"
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+            required
+          />
+        </div>
+        <div>
+          <label htmlFor="password">Password:</label>
+          <input
+            type="password"
+            id="password"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            required
+          />
+        </div>
+        <button type="submit">Login</button>
+      </form>
+    </Container>
+  );
+};
+
+export default Login;
